feat(profile): add images to profile page and refactor code to get images from Supabase storage
This commit adds two sections to the profile page: one for posts images and another for avatars images. The images are fetched from Supabase storage using the `getImagesUploadedByUser` function, which was refactored to accept a configuration object with the limit and bucket name. The `getPublicUrlsFromFiles` function was also added to get the public URLs of the images. The code was also refactored to use single quotes instead of double quotes.